requirements.txt
# Core dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.23

# Utilities
click==8.1.7

# Cursor IDE Integration
pyperclip==1.8.2
pywin32==306; sys_platform == "win32"